requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
twilio==8.10.0
openai==1.3.0
apscheduler==3.10.4
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.2
requests==2.31.0
